Refactored handle functions for navigate and location hooks
