+++ src/routes/details/Details.tsx
@@ -1,6 +1,7 @@
 import { Bullseye, Pagination, PaginationVariant, Spinner } from '@patternfly/react-core';
 import { Main } from '@redhat-cloud-services/frontend-components/Main';
 import type { Query } from 'api/queries';
+import { getQueryRoute } from 'api/queries';
 import type { Report } from 'api/reports/report';
 import { ReportPathsType } from 'api/reports/report';
 import type { AxiosError } from 'axios';
@@ -9,17 +10,12 @@ import React, { lazy, Suspense, useEffect, useState } from 'react';
 import type { WrappedComponentProps } from 'react-intl';
 import { injectIntl } from 'react-intl';
 import { useSelector } from 'react-redux';
+import { useLocation, useNavigate } from 'react-router-dom';
 import { ExportModal } from 'routes/components/export';
 import { PageHeading } from 'routes/components/page-heading';
 import { DateRangeType } from 'routes/utils/dateRange';
-import {
-  getRouteForQuery,
-  handleOnFilterAdded,
-  handleOnFilterRemoved,
-  handleOnPerPageSelect,
-  handleOnSetPage,
-  handleOnSort,
-} from 'routes/utils/history';
+import type { Filter } from 'routes/utils/filter';
+import { addFilterToQuery, removeFilterFromQuery } from 'routes/utils/filter';
 import type { RootState } from 'store';
 import { FetchStatus } from 'store/common';
 import { reportSelectors } from 'store/reports';
@@ -31,7 +27,6 @@ import { DetailsFilterToolbar } from './DetailsFilterToolbar';
 import { DetailsHeaderToolbar } from './DetailsHeaderToolbar';
 import { DetailsTable } from './DetailsTable';
 import { getDateRangeType, getGroupByType, getSourcesOfSpendType, GroupByType, SourcesOfSpendType } from './types';
-import { useLocation, useNavigate } from 'react-router-dom';
 const Loading = lazy(() => import('routes/state/loading/Loading'));
 const NotAvailable = lazy(() => import('routes/state/not-available/NotAvailable'));
 
@@ -85,8 +80,8 @@ const Details: React.FC<DetailsProps> = ({ intl }) => {
         groupBy={groupBy}
         isExportDisabled={!(report && report.meta) || report.meta.count === 0}
         onExportClicked={handleOnExportModalOpen}
-        onFilterAdded={filter => handleOnFilterAdded(navigate, location, query, filter)}
-        onFilterRemoved={filter => handleOnFilterRemoved(navigate, location, query, filter)}
+        onFilterAdded={handleOnFilterAdded}
+        onFilterRemoved={handleOnFilterRemoved}
         pagination={getPagination()}
         query={query}
       />
@@ -117,8 +112,8 @@ const Details: React.FC<DetailsProps> = ({ intl }) => {
       <Pagination
         isCompact={!isBottom}
         itemCount={count as number}
-        onPerPageSelect={(event, perPage) => handleOnPerPageSelect(history, location, query, perPage)}
-        onSetPage={(event, pageNumber) => handleOnSetPage(history, location, query, report, pageNumber)}
+        onPerPageSelect={handleOnPerPageSelect}
+        onSetPage={handleOnSetPage}
         page={page}
         perPage={limit}
         titles={{
@@ -133,15 +128,24 @@ const Details: React.FC<DetailsProps> = ({ intl }) => {
     );
   };
 
+  const getRouteForQuery = (newQuery: Query, reset: boolean = false) => {
+    // Reset pagination
+    if (reset) {
+      newQuery.filter = {
+        ...newQuery.filter,
+        offset: 0,
+      };
+    }
+    return `${location.pathname}?${getQueryRoute(newQuery)}`;
+  };
+
   const getTable = () => {
     return (
       <DetailsTable
         endDate={endDate}
         groupBy={groupBy}
         isLoading={reportFetchStatus === FetchStatus.inProgress}
-        onSort={(sortType, isSortAscending, date: string) =>
-          handleOnSort(history, query, sortType, isSortAscending, date)
-        }
+        onSort={handleOnSort}
         query={query}
         report={report}
         secondaryGroupBy={secondaryGroupBy}
@@ -163,6 +167,16 @@ const Details: React.FC<DetailsProps> = ({ intl }) => {
     setIsExportModalOpen(true);
   };
 
+  const handleOnFilterAdded = (filter: Filter) => {
+    const newQuery = addFilterToQuery(query, filter);
+    navigate(getRouteForQuery(newQuery, true), { replace: true });
+  };
+
+  const handleOnFilterRemoved = (filter: Filter) => {
+    const newQuery = removeFilterFromQuery(query, filter);
+    navigate(getRouteForQuery(newQuery, true), { replace: true });
+  };
+
   const handleOnGroupBySelected = value => {
     setGroupBy(value, () => {
       const groupByKey: keyof Query['group_by'] = value as any;
@@ -174,12 +188,48 @@ const Details: React.FC<DetailsProps> = ({ intl }) => {
         },
         order_by: { cost: 'desc' },
       };
-      navigate(getRouteForQuery(location, newQuery, true), { replace: true });
+      navigate(getRouteForQuery(newQuery, true), { replace: true });
     });
+  };
+
+  const handleOnPerPageSelect = (event, perPage) => {
+    const newQuery = { ...JSON.parse(JSON.stringify(query)) };
+    newQuery.filter = {
+      ...query.filter,
+      limit: perPage,
+    };
+    const filteredQuery = getRouteForQuery(newQuery, true);
+    navigate(filteredQuery, { replace: true });
   };
 
   const handleOnSecondaryGroupBySelected = value => {
     setSecondaryGroupBy(value);
+  };
+
+  const handleOnSetPage = (event, pageNumber) => {
+    const limit =
+      report && report.meta && report.meta.filter && report.meta.filter.limit ? report.meta.filter.limit : 10;
+    const offset = pageNumber * limit - limit;
+
+    const newQuery = { ...JSON.parse(JSON.stringify(query)) };
+    newQuery.filter = {
+      ...query.filter,
+      offset,
+    };
+    const filteredQuery = getRouteForQuery(newQuery);
+    navigate(filteredQuery, { replace: true });
+  };
+
+  const handleOnSort = (sortType: string, isSortAscending: boolean, date: string = undefined) => {
+    const newQuery = { ...JSON.parse(JSON.stringify(query)) };
+    newQuery.order_by = {};
+    newQuery.order_by[sortType] = isSortAscending ? 'asc' : 'desc';
+
+    if (date) {
+      newQuery.order_by.date = date;
+    }
+    const filteredQuery = getRouteForQuery(newQuery);
+    navigate(filteredQuery, { replace: true });
   };
 
   const handleOnSourcesOfSpendSelected = value => {
